Move current read index tests to read package
